+++ d3dtools/__init__.py
@@ -2,7 +2,7 @@
 D3DTOOLS - A collection of tools for working with shapefiles and converting them for Delft3D modeling.
 """
 
-__version__ = '0.23.0'
+__version__ = '0.24.2'
 
 # Define all modules that should be exposed when using "from d3dtools import *"
 __all__ = [
@@ -25,6 +25,9 @@ __all__ = [
     'getfacez2',
     'fou2shp',
     'pliz2shp',
+    'pli2shp',
+    'pol2shp',
+    'xyz2shp',
     'rmgrid',
 ]
 
@@ -47,4 +50,7 @@ from . import getfacez2
 from . import getfacez
 from . import fou2shp
 from . import pliz2shp
+from . import pli2shp
+from . import pol2shp
+from . import xyz2shp
 from . import rmgrid